pyproject.toml
[project]
name = "myproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "altair<5",
    "fastapi>=0.128.0",
    "joblib>=1.5.3",
    "numpy>=2.4.1",
    "pandas>=3.0.0",
    "pillow>=12.1.0",
    "python-multipart>=0.0.22",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.0",
    "streamlit>=1.19.0",
    "uvicorn>=0.40.0",
]
